# perf: improve aggregate rules