--- a/bundle/regal/rules/custom/missing-metadata/missing_metadata.rego
+++ b/bundle/regal/rules/custom/missing-metadata/missing_metadata.rego
@@ -22,9 +22,6 @@ _package_annotated if input["package"].annotations
 
 _rule_annotations[rule_path] contains annotated if {
 	some rule in ast.public_rules_and_functions
-	every part in rule.head.ref {
-		not startswith(part.value, "_")
-	}
 
 	rule_path := concat(".", [ast.package_name, ast.ref_static_to_string(rule.head.ref)])
 	annotated := count(object.get(rule, "annotations", [])) > 0
@@ -91,13 +88,10 @@ aggregate_report contains violation if {
 
 	any_item := util.any_set_item(aggregates)
 
-	violation := result.fail(rego.metadata.chain(), {"location": object.union(
-		any_item.location,
-		{
-			"file": any_item.file,
-			"text": split(any_item.location.text, "\n")[0],
-		},
-	)})
+	violation := result.fail(rego.metadata.chain(), {"location": object.union(any_item.location, {
+		"file": any_item.file,
+		"text": split(any_item.location.text, "\n")[0],
+	})})
 }
 
 # METADATA