--- a/bundle/regal/rules/imports/prefer-package-imports/prefer_package_imports_test.rego
+++ b/bundle/regal/rules/imports/prefer-package-imports/prefer_package_imports_test.rego
@@ -15,32 +15,8 @@ test_aggregate_collects_imports_with_location if {
 	r == {{
 		"aggregate_data": {
 			"imports": [
-				{
-					"location": {
-						"col": 2,
-						"file": "p.rego",
-						"row": 4,
-						"end": {
-							"col": 8,
-							"row": 4,
-						},
-						"text": "\timport data.b",
-					},
-					"path": ["b"],
-				},
-				{
-					"location": {
-						"col": 2,
-						"file": "p.rego",
-						"row": 5,
-						"end": {
-							"col": 8,
-							"row": 5,
-						},
-						"text": "\timport data.c.d",
-					},
-					"path": ["c", "d"],
-				},
+				[["b"], "4:2:4:8"],
+				[["c", "d"], "5:2:5:8"],
 			],
 			"package_path": ["a"],
 		},
@@ -51,28 +27,37 @@ test_aggregate_collects_imports_with_location if {
 
 test_fail_aggregate_report_on_imported_rule if {
 	r := rule.aggregate_report with input.aggregate as {
-		{"aggregate_data": {
-			"package_path": ["a"],
-			"imports": [
-				{
-					# likely import of rule — should fail
-					"path": ["b", "c"],
-					"location": {"col": 1, "file": "policy.rego", "row": 3, "text": "import data.b.c"},
-				},
-				# import of package, should not fail
-				{"path": ["b"], "location": {"col": 1, "file": "policy.rego", "row": 4, "text": "import data.b"}},
-				# unresolved import, should not fail
-				{"path": ["c"], "location": {"col": 1, "file": "policy.rego", "row": 5, "text": "import data.c"}},
-			],
-		}},
-		{"aggregate_data": {"package_path": ["b"], "imports": []}},
+		{
+			"aggregate_data": {
+				"package_path": ["a"],
+				"imports": [
+					[["b", "c"], "3:1:3:8"], # likely import of rule — should fail
+					[["b"], "4:1:4:8"], # import of package, should not fail
+					[["c"], "5:1:5:8"], # unresolved import, should not fail
+				],
+			},
+			"aggregate_source": {"file": "policy.rego", "package_path": ["a"]},
+		},
+		{
+			"aggregate_data": {"package_path": ["b"], "imports": []},
+			"aggregate_source": {"file": "policy2.rego", "package_path": ["b"]},
+		},
 	}
 
 	r == {{
 		"category": "imports",
 		"description": "Prefer importing packages over rules",
 		"level": "error",
-		"location": {"col": 1, "file": "policy.rego", "row": 3, "text": "import data.b.c"},
+		"location": {
+			"file": "policy.rego",
+			"col": 1,
+			"row": 3,
+			"end": {
+				"col": 8,
+				"row": 3,
+			},
+			"text": "import data.b.c",
+		},
 		"related_resources": [{
 			"description": "documentation",
 			"ref": config.docs.resolve_url("$baseUrl/$category/prefer-package-imports", "imports"),
@@ -85,10 +70,7 @@ test_success_aggregate_report_on_import_with_matching_package if {
 	r := rule.aggregate_report with input.aggregate as {
 		{"aggregate_data": {
 			"package_path": ["a"],
-			"imports": [{
-				"path": ["b"],
-				"location": {"col": 1, "file": "policy.rego", "row": 3, "text": "import data.b"},
-			}],
+			"imports": [[["b"], "3:1:3:8"]],
 		}},
 		{"aggregate_data": {
 			"package_path": ["b"],
@@ -105,10 +87,7 @@ test_success_aggregate_report_on_import_with_unresolved_path if {
 	r := rule.aggregate_report with input.aggregate as {
 		{"aggregate_data": {
 			"package_path": ["a"],
-			"imports": [{
-				"path": ["foo"],
-				"location": {"col": 1, "file": "policy.rego", "row": 3, "text": "import data.b"},
-			}],
+			"imports": [[["b"], "3:1:3:8"]],
 		}},
 		{"aggregate_data": {
 			"package_path": ["bar"],
@@ -123,10 +102,7 @@ test_success_aggregate_report_ignored_import_path if {
 	aggregate := {
 		{"aggregate_data": {
 			"package_path": ["a"],
-			"imports": [{
-				"path": ["b", "c"],
-				"location": {"col": 1, "file": "policy.rego", "row": 3, "text": "import data.b.c"},
-			}],
+			"imports": [[["b", "c"], "3:1:3:8"]],
 		}},
 		{"aggregate_data": {
 			"package_path": ["b"],
@@ -154,19 +130,7 @@ test_aggregate_ignores_imports_of_regal_in_custom_rule if {
 
 	r == {{
 		"aggregate_data": {
-			"imports": [{
-				"location": {
-					"col": 2,
-					"file": "p.rego",
-					"row": 6,
-					"end": {
-						"col": 8,
-						"row": 6,
-					},
-					"text": "\timport data.a.b.c",
-				},
-				"path": ["a", "b", "c"],
-			}],
+			"imports": [[["a", "b", "c"], "6:2:6:8"]],
 			"package_path": ["custom", "regal", "rules", "foo", "bar"],
 		},
 		"aggregate_source": {"file": "p.rego", "package_path": ["custom", "regal", "rules", "foo", "bar"]},