Make aggregate rules more efficient by only aggregating data that's absolutely needed, and try to use more compact representations for the aggregated data, the same way just done for unresolved-reference.

```
1130275167 ns/op	3204268280 B/op	62150928 allocs/op
1166222041 ns/op	3189028336 B/op	61860532 allocs/op - circular-import: defer to_location to report
1107372792 ns/op	3191798600 B/op	61459275 allocs/op - circular-import: use compact array format for aggregates
1122124750 ns/op	3183487808 B/op	61315825 allocs/op - prefer-package-imports: same as above
1095128583 ns/op	3171781928 B/op	61097537 allocs/op - unresolved-import: same as above
```

<!--
Thank you for submitting a pull request to Regal!

If you're new to contributing to the project, some tips and pointers are provided in the
contributing](https://github.com/StyraInc/regal/blob/main/docs/CONTRIBUTING.md) docs. If you find anything missing, or
not made clear enough, that's a bug, and we'd appreciate hearing about it!

If you want to ask questions before submitting your PR, or want to discuss Regal in general, please feel free to join
us in the `#regal` channel in the [Styra Community Slack](https://inviter.co/styra).
-->